refactor: 스레드 sleep 제거
