+++ src/main/java/usw/suwiki/global/util/loadjson/JsonToDataTable.java
@@ -17,7 +17,7 @@ import usw.suwiki.domain.lecture.domain.repository.LectureRepository;
 @Service
 @Transactional
 @RequiredArgsConstructor
-// TODO refactor: Json - JsonToLectureForm 변환 책임 분리
+// TODO refactor: LectureService로 이동
 // TODO feat: 데이터 적재 과정을 admin API로 생성
 public class JsonToDataTable {
 
@@ -27,7 +27,7 @@ public class JsonToDataTable {
     // TODO style: 메서드명 변경
     // TODO refactor: throws -> try catch
     // TODO fix: 강의 장소-교시 컬럼 (place_schedule) 누락
-    public void toEntity(String path) throws IOException, ParseException, InterruptedException {
+    public void toEntity(String path) throws IOException, ParseException {
         Reader reader = new FileReader(path);
 
         JSONParser parser = new JSONParser();
@@ -40,7 +40,7 @@ public class JsonToDataTable {
                 JSONObject jsonObject = (JSONObject) jsonArray.get(i);
                 JSONLectureVO jsonLectureVO = new JSONLectureVO(jsonObject);
 
-                Optional<Lecture> optionalLecture = lectureRepository.verifyJsonLecture(
+                Optional<Lecture> optionalLecture = lectureRepository.verifyJsonLecture(    // TODO refactor: 메서드명
                         jsonLectureVO.getLectureName(),
                         jsonLectureVO.getProfessor(),
                         jsonLectureVO.getMajorType()
@@ -56,7 +56,6 @@ public class JsonToDataTable {
                 } else {
                     // TODO refactor: JsonToLectureForm -> Lecture 의존하도록
                     Lecture newLecture = jsonLectureVO.toEntity();
-                    Thread.sleep(1);    // TODO figure: 왜 필요하지?
                     lectureRepository.save(newLecture);
                 }
