+++ src/test/java/usw/suwiki/lectureutill/InsertNewLectureTest.java
@@ -1,15 +1,16 @@
 package usw.suwiki.lectureutill;
 
+import java.io.IOException;
 import org.json.simple.parser.ParseException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 import usw.suwiki.domain.lecture.domain.repository.LectureRepository;
-import usw.suwiki.global.util.JsonToDataTable;
+import usw.suwiki.global.util.loadjson.JsonToDataTable;
 
-import java.io.IOException;
-
-// @ActiveProfiles("test")
+// TODO: 다른 테스트 코드들에도 "test" 프로파일 설정 혹은 default 프로파일로 변경. ActiveProfile 필요성 고민
+@ActiveProfiles("test")
 @SpringBootTest
 class InsertNewLectureTest {
 
@@ -20,7 +21,7 @@ class InsertNewLectureTest {
     LectureRepository lectureRepository;
 
     @Test
-    void insertNewLecture() throws IOException, ParseException, InterruptedException {
+    void insertNewLecture() throws IOException, ParseException {
         jsonToDataTable.toEntity("src/main/resources/USW_2023_2.json");
     }
 }
